requirements.txt
# Core Data Processing
pyspark==3.4.1
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1

# Machine Learning
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2

# Database
psycopg2-binary==2.9.6
SQLAlchemy==2.0.19

# Streaming (Optional)
kafka-python==2.0.2

# Orchestration (Optional)
apache-airflow==2.6.3
apache-airflow-providers-postgres==5.5.1

# ML Ops
mlflow==2.5.0

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
click==8.1.6

# Visualization & Analysis
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Jupyter (for notebooks)
jupyter==1.0.0
ipykernel==6.25.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.7.0
flake8==6.1.0
mypy==1.5.0

# Logging
loguru==0.7.0
